up to septillion

+++ Main/Models/NumberDictionary.cs
@@ -1,11 +1,12 @@
 using System;
 using System.Collections.Generic;
+using System.Numerics;
 
 namespace NumberWordsDictionary.Models
 {
   public class NumberDictionary
   {
-    private static IDictionary<int,string> dict = new Dictionary<int, string>()
+    private static IDictionary<BigInteger,string> dict = new Dictionary<BigInteger, string>()
     {
       {0,""},{1,"One"},{2,"Two"},{3,"Three"},{4,"Four"},{5,"Five"},{6,"Six"},{7,"Seven"},
       {8,"Eight"},{9,"Nine"},{10,"Ten"},{11,"Eleven"},{12,"Twelve"},{13,"Thirteen"},{14,"Fourteen"},
@@ -13,7 +14,7 @@ namespace NumberWordsDictionary.Models
       {30,"Thirty"},{40,"Fourty"},{50,"Fifty"},{60,"Sixty"},{70,"Seventy"},{80,"Eighty"},{90,"Ninety"},{100,"Hundred"}
     };
 
-    public static string GetValue(int i)
+    public static string GetValue(BigInteger i)
     {
       if(dict.ContainsKey(i))
       {
